backend/requirements.txt
# Backend Requirements
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
